% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Useful_functions.R
\name{freedman_diaconis_binwidth_log}
\alias{freedman_diaconis_binwidth_log}
\title{Function to calculate bin width but in a log axis}
\usage{
freedman_diaconis_binwidth_log(x)
}
\arguments{
\item{x}{(vector) data to be plotted as histogram (eg. column of dataframe)}
}
\description{
Function to calculate bin width but in a log axis
}
