% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Useful_functions.R
\name{list_package_functions}
\alias{list_package_functions}
\title{List exported functions of an installed package}
\usage{
list_package_functions()
}
\value{
A character vector of function names
}
\description{
List exported functions of an installed package
}
