% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Plot_IRAS_file.R
\name{plot_IRAS}
\alias{plot_IRAS}
\title{Plot ActiWiz IRAS vs calculated IRAS}
\arguments{
\item{filename}{Where is the items file located and what's the name}

\item{filetype}{"items" or "packages" depending on the file you plot}
}
\description{
Function to plot (almost) the relationship of IRAS
}
