% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Plot_items_file.R
\name{plot_items}
\alias{plot_items}
\title{Plot all items characteristics}
\arguments{
\item{filename}{Where is the items file located and what's the name}
}
\description{
Function to plot (almost) all item characteristics in one grid
}
