% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Plot_scenarios_file.R
\name{plot_scenarios}
\alias{plot_scenarios}
\title{Plot all scenario parameters}
\arguments{
\item{filename}{Where is the scenarios file located and what's the name}
}
\description{
Function to plot (almost) all scenario parameters in one grid
}
