% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Theme.R
\docType{data}
\name{theme_barplot_axes}
\alias{theme_barplot_axes}
\title{Theme that edits both axes so they look nice for a barplot
No x axis and no title for y axis. Also no grid.}
\format{
An object of class \code{theme} (inherits from \code{gg}) of length 6.
}
\usage{
theme_barplot_axes
}
\description{
Theme that edits both axes so they look nice for a barplot
No x axis and no title for y axis. Also no grid.
}
\keyword{datasets}
