% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Theme.R
\docType{data}
\name{theme_no_x_axis}
\alias{theme_no_x_axis}
\title{Theme that deletes the x axis of a plot}
\format{
An object of class \code{theme} (inherits from \code{gg}) of length 4.
}
\usage{
theme_no_x_axis
}
\description{
Theme that deletes the x axis of a plot
}
\keyword{datasets}
